Reorder test classes
