+++ owncloudApp/src/androidTest/java/com/owncloud/android/shares/data/OCLocalShareDataSourceTest.kt
@@ -17,13 +17,15 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.owncloud.android.shares.data.datasources
+package com.owncloud.android.shares.data
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.MutableLiveData
 import androidx.test.platform.app.InstrumentationRegistry
 import com.owncloud.android.db.OwncloudDatabase
 import com.owncloud.android.lib.resources.shares.ShareType
+import com.owncloud.android.shares.data.datasources.OCLocalSharesDataSource
+import com.owncloud.android.shares.data.datasources.OCShareDao
 import com.owncloud.android.shares.domain.OCShare
 import com.owncloud.android.utils.LiveDataTestUtil.getValue
 import com.owncloud.android.utils.TestUtil
@@ -83,7 +85,8 @@ class OCLocalDataSourceTest {
         `when`(db.shareDao()).thenReturn(ocSharesDao)
 
         val context = InstrumentationRegistry.getInstrumentation().targetContext
-        ocLocalSharesDataSource = OCLocalSharesDataSource(context, ocSharesDao)
+        ocLocalSharesDataSource =
+            OCLocalSharesDataSource(context, ocSharesDao)
     }
