+++ owncloudApp/src/androidTest/java/com/owncloud/android/shares/data/OCShareDaoTest.kt
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.owncloud.android.shares.db
+package com.owncloud.android.shares.data
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
