+++ owncloudApp/src/androidTest/java/com/owncloud/android/shares/presentation/ui/ShareFileFragmentTest.kt
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.owncloud.android.shares.ui
+package com.owncloud.android.shares.presentation.ui
 
 import android.accounts.Account
 import androidx.test.espresso.Espresso.onView
