+++ owncloudApp/src/androidTest/java/com/owncloud/android/shares/presentation/ui/SharesContentProviderTest.kt
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.owncloud.android.shares.provider
+package com.owncloud.android.shares.presentation.ui
 
 import android.content.ContentResolver
 import android.content.ContentValues
